mem: still reachable memleaks
uthash never frees an entry on call to HASH_DEL therefore requiring caller to handle. refactored add_transaction to check if index exists for entry and if not it adds to hash table. change from autoincremented int to uint256 hash to int for more unique id field in new_transaction().