net: basic networking capabilities
new optional feature (default enabled, to disable run ./configure --disable-net) includes minimum dependencies needed in order to interact with the dogecoin network. in addition to that option users now have control over disabling the command line tool such by running ./configure --disable-tools. it is enabled by default as well. the pertinent information regarding build options are in Makefile.am, configure.ac and test/unittester.c.
this feature is comprised of 3 building blocks made up in the form of block, which contains definition of dogecoin_block_header and methods needed to interact with that structure, net which is the bridge between event2\event.h and contains definitions for dogecoin_node_group, node_connectionstate and dogecoin_node structures and methods to properly configure and interact with external peers residing on the dogecoin network and finally protocol which contains structures that define how to interact with external peers on the network.
each of the aforementioned building blocks that enable networking functionality have corresponding unit tests.