transaction: adds simplified transaction signing function
until we've implemented some sort of utxo cache such as in the spv-wallet branch with changes slated for a later release, we'll be unable to fetch utxo amounts. therefore at a minimum users will need to provide their own utxo amounts within an array in sequential order to bulk sign inputs. this functionality will of course be available soon, however for the interim this function provides a more simplified approach and allows signing without having to pass in the raw hexadecimal string itself as it's retrieved from memory and loaded internally. added a duplicated transaction signing test in transaction_tests.c that shows an example of potential usage by signing both inputs within a single function call using an array containing the previously known utxo amounts..