Renamed HD key conversion functions with more sensible names.
The old name `generateDerivedHDPubKey` is misleading - it doesn't
generate anything new, it just converts the HD pubkey to P2PKH format.
It also works for any HD child address, so added that to docs.

Backwards compatibility: there's a 1-line wrapper function with the
old name, to maintain binary compatibility.

Also renamed my `generateDerivedHDPrivKeyWIF` (which was named that
way to be consistent!) - it just converts the HD privkey to a
non-extended non-HD private key (EC Privkey)

The EC Privkey is required to sign transactions, since you can't
sign a transaction using an extended HD key, the API doesn't support it.