Improve HD address derivation
The existing functions generated keys, not addresses. This code extracts
common behavior to be shared between these functions and adds a function
that return addresses (using the name of the previous function) and adds
a new function name (using the body of the previous function) that
returns HD public or private keys.

Addresses GH #152.